move EvalSpec into test sub-module
