+++ vespalib/src/vespa/vespalib/eval/test/CMakeLists.txt
@@ -1,0 +1,6 @@
+# Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+vespa_add_library(vespalib_vespalib_eval_test OBJECT
+    SOURCES
+    eval_spec.cpp
+    DEPENDS
+)
